[Lab 02] Ex - Outsourced SpotAdapter in own file
