+++ app/src/main/res/layout/card_spot.xml
@@ -11,17 +11,17 @@
         android:padding="16dp">
 
         <TextView
-            android:id="@+id/spotTitle"
+            android:id="@+id/spotTitle_Card"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignParentTop="true"
             android:textSize="30sp" />
 
         <TextView
-            android:id="@+id/spotDescription"
+            android:id="@+id/spotDescription_Card"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_below="@id/spotTitle" />
+            android:layout_below="@id/spotTitle_Card" />
 
     </RelativeLayout>
 </androidx.cardview.widget.CardView>
